refactor(screamingface-engine): fold ifeval onto the shared scored spine
Selection, row decode, and the failure ladder now come from spine/scored.py;
ifeval keeps only its deterministic grade_case hook, its published accuracy
scorer, and its failure wording. ifeval/aggregate.py (442 lines) is deleted;
vendor/ and the rendered expression are untouched.

The spine's exam reduction becomes a per-board scorer parameter (rubric boards
bind exam_scorer(mean) at their call sites, byte-identical output), and gains
one optional board-owned missing-row hook so ifeval's golden-pinned
collected-row failure wording stays byte-identical and the OME-981
candidate-vs-grading boundary decision stays open.

All four golden replays green (ifeval 0.9184 incl. its failure pin,
gdpval-text, healthbench-worst30, draco-3pass); all 80 pre-fold ifeval unit
tests pass with assertions untouched.

Refs: OME-1101
